feat: 충전 테스트 추가
