+++ src/test/java/kr/hhplus/be/server/point/PointServiceTest.java
@@ -1,8 +1,11 @@
 package kr.hhplus.be.server.point;
 
-import kr.hhplus.be.server.point.domain.model.PointBalance;
-import kr.hhplus.be.server.point.infrastructure.persistence.mapper.PointMapper;
 import kr.hhplus.be.server.point.application.service.PointService;
+import kr.hhplus.be.server.point.domain.model.Point;
+import kr.hhplus.be.server.point.domain.model.PointBalance;
+import kr.hhplus.be.server.point.domain.model.enums.PointType;
+import kr.hhplus.be.server.point.domain.repository.PointRepository;
+import kr.hhplus.be.server.point.infrastructure.persistence.mapper.PointMapper;
 import kr.hhplus.be.server.user.application.service.UserService;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -10,15 +13,19 @@ import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.context.SpringBootTest;
 
-import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-@ExtendWith(MockitoExtension.class)
+@SpringBootTest
 public class PointServiceTest {
 
-    @InjectMocks
+    @Autowired
     private PointService pointService;
+
+    @Autowired
+    private PointRepository pointRepository;
 
     @Mock
     private PointMapper pointMapper;
@@ -26,15 +33,22 @@ public class PointServiceTest {
     @Mock
     private UserService userService;
 
-    @DisplayName("쿼리 조회 결과가 null인 경우, 0원을 반환한다.")
+    @DisplayName("포인트를 10000원 충전한다.")
     @Test
-    void ifNullThenReturnZeroBalance() {
-        //given
-        final int userNo = 0;
+    void chargeTest() {
+        final int userNo = 1;
+        final long chargeAmount = 10000;
+        PointBalance balance = pointService.selectBalance(userNo);
 
-       //when then
-        assertThatThrownBy(() -> pointService.selectBalance(userNo));
+        Point point = Point.builder()
+                .type(PointType.CHARGE)
+                .amount(chargeAmount)
+                .balance(balance.getBalance() + chargeAmount)
+                .userNo(userNo)
+        .build();
+
+        // 포인트 충전 이력 생성
+        pointRepository.insertPointHist(point);
     }
-
 
 }
